+++ src/Command/PluginAssetsTrait.php
@@ -18,7 +18,7 @@ namespace Cake\Command;
 
 use Cake\Core\Configure;
 use Cake\Core\Plugin;
-use Cake\Filesystem\Filesystem;
+use Cake\Utility\Filesystem;
 use Cake\Utility\Inflector;
 
 /**
