+++ src/Console/composer.json
@@ -26,7 +26,6 @@
         "php": ">=8.0",
         "cakephp/core": "^5.0",
         "cakephp/event": "^5.0",
-        "cakephp/filesystem": "^5.0",
         "cakephp/log": "^5.0",
         "cakephp/utility": "^5.0"
     },
