--- src/Filesystem/composer.json
@@ -1,34 +1,0 @@
-{
-    "name": "cakephp/filesystem",
-    "description": "CakePHP filesystem convenience classes to help you work with files and folders.",
-    "type": "library",
-    "keywords": [
-        "cakephp",
-        "filesystem",
-        "files",
-        "folders"
-    ],
-    "homepage": "https://cakephp.org",
-    "license": "MIT",
-    "authors": [
-        {
-            "name": "CakePHP Community",
-            "homepage": "https://github.com/cakephp/filesystem/graphs/contributors"
-        }
-    ],
-    "support": {
-        "issues": "https://github.com/cakephp/cakephp/issues",
-        "forum": "https://stackoverflow.com/tags/cakephp",
-        "irc": "irc://irc.freenode.org/cakephp",
-        "source": "https://github.com/cakephp/filesystem"
-    },
-    "require": {
-        "php": ">=8.0",
-        "cakephp/core": "^5.0"
-    },
-    "autoload": {
-        "psr-4": {
-            "Cake\\Filesystem\\": "."
-        }
-    }
-}
