+++ src/Utility/Filesystem.php
@@ -14,7 +14,7 @@ declare(strict_types=1);
  * @since         4.0.0
  * @license       https://opensource.org/licenses/mit-license.php MIT License
  */
-namespace Cake\Filesystem;
+namespace Cake\Utility;
 
 use Cake\Core\Exception\CakeException;
 use CallbackFilterIterator;
